Push the wait-for-job to the conftest
